Fix race condition in `SavedTerm::load(env)`.
I originally intended for a `SavedTerm` to travel with the `OwnedEnv`
where it was saved. They might both be moved to another thread, for
example.

Unfortunately, the code I actually wrote is more permissive than that!
Without this patch, a `SavedTerm` can be moved to another thread and
loaded into an environment there, *without* also moving the original
`OwnedEnv` to that thread.

`SavedTerm::load` checks that the source `OwnedEnv` still exists, but
it's a point-in-time check; nothing prevents the `OwnedEnv` from being
destroyed in the original thread while the `SavedTerm::load` thread is
copying the value.

How to fix? Keeping the environment around after the `OwnedEnv` is gone
would be possible, but too weird. So we tighten up the contract and
panic if you try to load a `SavedTerm` into an environment other than
the `OwnedEnv` where it was created.