Add rustler attributes for conditional compilation
This commit adds attributes for conditional code generation of encode and
decode. Users can now choose if they want both encode and decode
functions to be derived, or only one of those functions.

For example, for the following struct `A`, only encode is derived:

```rust
struct A<'a> {
    my_string: &'a str,
}
```

Additionally, this commit also refactors how the module for Elixir
structs or the tag for Erlang records can be set. In addition to the
format `#[module = "A.B.C"]` or `#[tag = "a"]`, the `rustler` attribute
may also be used:

```rust
struct A<'a> {
    my_string: &'a str,
}
```