Don't call cargo when skip_compilation? is true
This fixes a regression from #386 which made rustler_mix call
`cargo metadata` even when skip_compilation? is set to true,
breaking setups in which the rust library is built separately, e.g.
in separate docker stage with proper rust versioning and in which
the elixir compilation stage doesn't have cargo.

This fix assumes that the external_resources attribute is
irrelevant when skip_compilation? is set to true.