Rewrite build of NIF APIs in Rust
This changes the way NIF APIs are generated by rewriting the build
script in Rust.  The idea is to remove the dependency of having
Erlang installed.

The way we do now is to read the `RUSTLER_NIF_VERSION` env var in
the `build.rs` files and use that version to generate the NIF API file
before compiling the rest of Rustler's code. If that environment var
is not available, then we try to read from the installed Erlang. If the
build machine doesn't have Erlang, then we use the latest supported
version and write a warning to the stderr of the scripts (you can find
the messages at `target/debug/build/<crate-name>-<hash>/stderr`).

Now all crates compiled directly from a project using Rustler will pass
the correct NIF version in the env var `RUSTLER_NIF_VERSION`. See the
Rustler compiler in Elixir for details.

This is a step to make easier the cross-compilation of Rustler
crates to different NIF versions without having to rely on the current
installed OTP.

We also don't need the precompiled snippets because we can always
generate these files now.