Cache Mix.env() to trigger recompilation
When using MIX_ENV to configure the build profile, changing `MIX_ENV` in
a subsequent run needs to trigger the mechanism for recompilation to
determine if that mode was build. For this, we need to cache value of
`Mix.env()` and detect when that changes. If it changes, we write the
cache, `@external_resource` can pick that up.