Extend load_from to allow loading from an absolute path
`load_from` already has a format to specify a path relative to an OTP
application. This commit extends the type for `load_from` to also allow
setting an absolute path. Note that no special check for the validity of
the path is being done, as we do not know beforehand what the target
operating system looks like. Setting this configuration correctly is
thus the user's responsibility.