Implement simple allocator using enif_{alloc,free} (#580)
Since we can only expect a 64bit alignment from the Erlang allocator, we
have to ensure larger alignments by overallocating. The allocator for
this case behaves the same way as the zigler "large beam allocator", see
https://github.com/E-xyza/zigler/blob/main/priv/beam/allocator.zig.

If the alignment is greater than 8, we allocate enough memory to store
an additional pointer. The result of the initial allocation is then
written immediately before the aligned pointer, which is returned from
the allocator. When deallocating, we can retrieve the original pointer
and pass it on to `enif_free`.